A little of clean up
